fix: Resolve clippy warnings in CI and simplify README

+++ rtk/tests/common/mod.rs
@@ -1,5 +1,4 @@
-/// Shared test utilities — imported by integration tests via `mod common`.
-
+//! Shared test utilities — imported by integration tests via `mod common`.
 /// Counts whitespace-delimited tokens (proxy for LLM token count).
 pub fn count_tokens(text: &str) -> usize {
     text.split_whitespace().count()
@@ -12,6 +11,6 @@ pub fn token_savings(original: &str, filtered: &str) -> (f64, bool) {
     if orig == 0 {
         return (0.0, false);
     }
-    let savings = 1.0 - (filt as f64 / orig as f64);
-    (savings * 100.0, savings >= 0.60)
+    let savings = 1.0 - (f64::from(filt as u32) / f64::from(orig as u32));
+    (savings * 100.0, savings >= 0.6)
 }
